requirements.txt
torch>=2.0.0
torchvision
transformers>=4.37.0
qwen-vl-utils
pillow
numpy
accelerate
fastapi>=0.95.0
uvicorn[standard]>=0.18.0
python-multipart>=0.0.5
